Refac(js): js파일 모듈화
