+++ README.md
@@ -1,5 +1,16 @@
 # javascript-baseball-precourse
 
+## 최종 구현한 기능
+
+1. 사용자 인터페이스 구현
+2. 3자리수 타겟 생성
+3. 유효한 사용자 입력 받기
+4. 사용자 입력과 타겟 비교해서 점수 매기기
+5. 결과를 사용자에게 공지
+6. 게임 재시작 버튼
+9. 모듈화
+
+
 ### 과제 진행 요구 사항
 
 * 미션은 숫자 야구 저장소를 포크하고 클론하는 것으로 시작한다.
